Seperate huge selectors.ts file into child modules
