+++ webpack/resources/selectors_for_indexing.ts
@@ -1,4 +1,12 @@
-import { TaggedResource } from "./tagged_resources";
+import {
+  TaggedResource,
+  TaggedToolSlotPointer,
+  TaggedSequence,
+  TaggedRegimen,
+  TaggedFarmEvent,
+  TaggedTool,
+  TaggedPoint
+} from "./tagged_resources";
 import { CowardlyDictionary } from "../util";
 import { ResourceIndex } from "./interfaces";
 import { assertUuid } from "./selectors";
@@ -10,16 +18,14 @@ interface Indexer<T extends TaggedResource> {
   (index: ResourceIndex): IndexLookupDictionary<T>;
 }
 
-interface MapperFn<T extends TaggedResource> {
-  (item: T): T | undefined;
-}
+interface MapperFn<T extends TaggedResource> { (item: T): T | undefined; }
 
 /** Build a function,
- *    that returns a function,
- *      that returns a dictionary,
- *        that contains TaggedResource of kind T
- *          that uses the resource's id as the dictionary key.
- * */
+*    that returns a function,
+*      that returns a dictionary,
+*        that contains TaggedResource of kind T
+*          that uses the resource's id as the dictionary key.
+* */
 export const buildIndexer =
   <T extends TaggedResource>(kind: T["kind"], mapper?: MapperFn<T>): Indexer<T> => {
     return function (index: ResourceIndex, ) {
@@ -40,3 +46,17 @@ export const buildIndexer =
       return output;
     };
   };
+
+const slotMapper = (i: TaggedPoint): TaggedToolSlotPointer | undefined => {
+  if (i.kind == "Point" && (i.body.pointer_type === "ToolSlot")) {
+    return i as TaggedToolSlotPointer;
+  }
+  return undefined;
+};
+
+export const indexBySlotId =
+  buildIndexer<TaggedToolSlotPointer>("Point", slotMapper);
+export const indexSequenceById = buildIndexer<TaggedSequence>("Sequence");
+export const indexRegimenById = buildIndexer<TaggedRegimen>("Regimen");
+export const indexFarmEventById = buildIndexer<TaggedFarmEvent>("FarmEvent");
+export const indexByToolId = buildIndexer<TaggedTool>("Tool");
